Fixed signed comparison warning

+++ inc/umocktypes.h
@@ -50,7 +50,7 @@ extern "C" {
         } \
         else \
         { \
-            if ((int)*value < sizeof(C2(enum_name,_strings)) / sizeof(C2(enum_name,_strings)[0])) \
+            if ((int)*value < (int)(sizeof(C2(enum_name,_strings)) / sizeof(C2(enum_name,_strings)[0]))) \
             { \
                 size_t length = strlen(C2(enum_name_, strings)[*value]); \
                 if (length == 0) \
